// types/contracts/idea3/IdeaSBT.ts
export declare namespace IdeaSBT {
  export type IdeaStructStruct = {
    id: PromiseOrValue<BigNumberish>;
    title: PromiseOrValue<string>;
    desc: PromiseOrValue<string>;
    markdown: PromiseOrValue<string>;
    submitter: PromiseOrValue<string>;
    createAt: PromiseOrValue<BigNumberish>;
    updateAt: PromiseOrValue<BigNumberish>;
    approved: PromiseOrValue<boolean>;
    submitterHandle: PromiseOrValue<string>;
  };

  export type IdeaStructStructOutput = [
    BigNumber,
    string,
    string,
    string,
    string,
    BigNumber,
    BigNumber,
    boolean,
    string
  ] & {
    id: BigNumber;
    title: string;
    desc: string;
    markdown: string;
    submitter: string;
    createAt: BigNumber;
    updateAt: BigNumber;
    approved: boolean;
    submitterHandle: string;
  };
}